chore: Rename in invitation to CreateInvitation
closes #6
