+++ components/invitation/CreateInvitation.stories.js
@@ -4,14 +4,14 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 import { storiesOf } from '@storybook/vue'
-import invitation from './invitation'
+import createInvitation from './CreateInvitation'
 
 
 storiesOf('invitation', module)
-    .add('normal', () => ({
-        components: { invitation },
+    .add('create', () => ({
+        components: { createInvitation },
         template:
             `
-      <invitation/>
+      <createInvitation/>
     `
     }))
