+++ components/invitation/index.js
@@ -3,4 +3,13 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 
 SPDX-License-Identifier: Apache-2.0
 */
-export * from './invitation.vue'
+
+import CreateInvitation from "./CreateInvitation.vue";
+
+export default {
+    install(Vue, options) {
+        // Let's register our component globally
+        // https://vuejs.org/v2/guide/components-registration.html
+        Vue.component("create-invitation", CreateInvitation);
+    }
+};
