+++ components/invitation/package.json
@@ -1,7 +1,8 @@
 {
   "name": "@trustbloc/invitation",
-  "version": "0.0.1",
+  "version": "0.0.6",
   "publishConfig": {
     "access": "public"
-  }
+  },
+  "main": "./index.js"
 }
